requirements.txt
# 交易所API
ccxt==3.1.54
python-binance==1.0.17
okx==2.1.2

# 数据分析
pandas==2.0.3
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0

# 技术分析
ta-lib-python>=0.4.24
pandas-ta==0.3.14b0

# 机器学习
scikit-learn>=1.2.0

# 社交媒体分析
selenium==4.18.1
webdriver-manager==4.0.1
textblob==0.17.1  # 情感分析
nltk==3.8.1

# AI决策系统
requests>=2.28.2
json5>=0.9.10
python-dotenv>=1.0.0

# 工具
tqdm>=4.64.1
colorama>=0.4.6 